data-source-service Split the interfaces from the service into seperate models
