+++ angular/src/app/models/data-source-service.ts
@@ -1,0 +1,5 @@
+export interface IDataSourceService {
+    createDataSource(): void;
+    getDataSource(source: string): void;
+    updateDataSource(source: string, value: any): void
+}
